[Librarian] Regenerated @ affbf0251ae3f327c6613374357247a3417b863e

+++ src/Twilio/Rest/Video/V1/RecordingResource.cs
@@ -414,7 +414,7 @@ namespace Twilio.Rest.Video.V1
         /// The size
         /// </summary>
         [JsonProperty("size")]
-        public int? Size { get; private set; }
+        public long? Size { get; private set; }
         /// <summary>
         /// The url
         /// </summary>
